src/main/java/com/Laxmi/financeManager/service/TransactionService.java
package com.Laxmi.financeManager.service;
import com.Laxmi.financeManager.entity.Transaction;
import com.Laxmi.financeManager.repository.TransactionRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class TransactionService {
    @Autowired
    private TransactionRepository transactionRepository;

    public Transaction saveTransaction(Transaction transaction) {
        return transactionRepository.save(transaction);
    }


    public List<Transaction> getAllTransactions() {
        return transactionRepository.findAll();
    }
}
